requirements.txt
flask
flask-smorest
python-dotenv
flask-sqlalchemy
flask-migrate
flask-socketio
eventlet
flask-jwt-extended
passlib